refactor: 코드리뷰 반영 validaterUser 책임 Question으로 이관
