+++ src/main/java/nextstep/users/domain/NsUser.java
@@ -118,12 +118,6 @@ public class NsUser {
         return false;
     }
 
-    public void validateUser(NsUser loginUser) throws CannotDeleteException {
-        if (!this.equals(loginUser)) {
-            throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
-        }
-    }
-
     private static class GuestNsUser extends NsUser {
         @Override
         public boolean isGuestUser() {
