+++ src/test/java/nextstep/users/domain/NsUserTest.java
@@ -12,19 +12,5 @@ public class NsUserTest {
     public static final NsUser SANJIGI = new NsUser(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
 
 
-    @Test
-    public void isOwner테스트() throws CannotDeleteException {
-        assertThatThrownBy(() ->
-        {
-            JAVAJIGI.validateUser(SANJIGI);
-        }).isInstanceOf(CannotDeleteException.class);
-    }
 
-    @Test
-    public void isOwner성공테스트() throws CannotDeleteException {
-        assertDoesNotThrow(() ->
-        {
-            JAVAJIGI.validateUser(JAVAJIGI);
-        });
-    }
 }
